finished EditDeck styling, edit card feature, level badges and level calc

+++ src/components/DeckDetails/CardsList/CardsList.js
@@ -1,28 +1,39 @@
 import { useMutation } from '@apollo/client';
-import { CREATE_CARD, DELETE_CARD, GET_DECK } from 'queries/queries';
+import PopupMessage from 'components/common/PopupMessage/PopupMessage';
+import { CREATE_CARD, DELETE_CARD, EDIT_CARD, GET_DECK } from 'queries/queries';
 import { useState } from 'react';
 
 const CardsList = ({ deckId, cards, editable }) => {
   const [front, setFront] = useState('');
   const [back, setBack] = useState('');
+  const [editRow, setEditRow] = useState({
+    index: -1,
+    front: '',
+    back: '',
+  });
 
   const options = {
     refetchQueries: [{ query: GET_DECK, variables: { id: deckId } }],
   };
 
-  const [createCard] = useMutation(CREATE_CARD, options);
+  const [createCard, { error: createError }] = useMutation(
+    CREATE_CARD,
+    options
+  );
+  const [editCard, { error: editError }] = useMutation(EDIT_CARD, options);
   const [deleteCard] = useMutation(DELETE_CARD, options);
 
   const handleCreateCard = async () => {
+    if (!front || !back) return;
+
     try {
       await createCard({
         variables: {
-          deckId: deckId,
-          front: front,
-          back: back,
+          deckId,
+          front,
+          back,
         },
       });
-
       setFront('');
       setBack('');
     } catch (err) {
@@ -31,12 +42,33 @@ const CardsList = ({ deckId, cards, editable }) => {
   };
 
   const handleDeleteCard = async (id) => {
+    if (window.confirm('Are you sure you want to delete this card?')) {
+      try {
+        await deleteCard({
+          variables: {
+            id,
+          },
+        });
+      } catch (err) {
+        console.log(err);
+      }
+    }
+  };
+
+  const handleEditCard = async (id) => {
+    const { front, back } = editRow;
+
+    if (!front || !back) return;
+
     try {
-      await deleteCard({
+      await editCard({
         variables: {
-          id: id,
+          id,
+          front,
+          back,
         },
       });
+      setEditRow({ index: -1, front: '', back: '' });
     } catch (err) {
       console.log(err);
     }
@@ -44,6 +76,10 @@ const CardsList = ({ deckId, cards, editable }) => {
 
   return (
     <div id="cards-list">
+      {createError && (
+        <PopupMessage message={createError.message} type="error" />
+      )}
+      {editError && <PopupMessage message={editError.message} type="error" />}
       <h3>Cards ({cards.length})</h3>
       <table>
         <thead>
@@ -55,23 +91,78 @@ const CardsList = ({ deckId, cards, editable }) => {
           </tr>
         </thead>
         <tbody>
-          {cards.map((card, index) => (
-            <tr key={index}>
-              <td>#{index + 1}</td>
-              <td>{card.front}</td>
-              <td>{card.back}</td>
-              {editable && (
+          {cards.map((card, index) =>
+            editRow.index === index ? (
+              <tr key={index}>
+                <td>#{index + 1}</td>
+                <td>
+                  <input
+                    type="text"
+                    value={editRow.front}
+                    onChange={(e) =>
+                      setEditRow({
+                        index: index,
+                        front: e.target.value,
+                        back: editRow.back,
+                      })
+                    }
+                    onKeyPress={(e) =>
+                      e.key === 'Enter' && handleEditCard(card.id)
+                    }
+                  />
+                </td>
+                <td>
+                  <input
+                    type="text"
+                    value={editRow.back}
+                    onChange={(e) =>
+                      setEditRow({
+                        index: index,
+                        front: editRow.front,
+                        back: e.target.value,
+                      })
+                    }
+                    onKeyPress={(e) =>
+                      e.key === 'Enter' && handleEditCard(card.id)
+                    }
+                  />
+                </td>
                 <td>
                   <span
-                    className="material-icons-outlined icon-btn"
-                    onClick={() => handleDeleteCard(card.id)}
+                    className="material-icons icon-btn"
+                    onClick={() => handleEditCard(card.id)}
                   >
-                    remove_circle
+                    check_circle
                   </span>
                 </td>
-              )}
-            </tr>
-          ))}
+              </tr>
+            ) : (
+              <tr
+                key={index}
+                onDoubleClick={() =>
+                  setEditRow({
+                    index: index,
+                    front: card.front,
+                    back: card.back,
+                  })
+                }
+              >
+                <td>#{index + 1}</td>
+                <td>{card.front}</td>
+                <td>{card.back}</td>
+                {editable && (
+                  <td>
+                    <span
+                      className="material-icons-outlined icon-btn"
+                      onClick={() => handleDeleteCard(card.id)}
+                    >
+                      remove_circle
+                    </span>
+                  </td>
+                )}
+              </tr>
+            )
+          )}
           {editable && (
             <tr>
               <td></td>
@@ -80,7 +171,7 @@ const CardsList = ({ deckId, cards, editable }) => {
                   type="text"
                   value={front}
                   onChange={(e) => setFront(e.target.value)}
-                  required
+                  onKeyPress={(e) => e.key === 'Enter' && handleCreateCard()}
                 />
               </td>
               <td>
@@ -88,7 +179,7 @@ const CardsList = ({ deckId, cards, editable }) => {
                   type="text"
                   value={back}
                   onChange={(e) => setBack(e.target.value)}
-                  required
+                  onKeyPress={(e) => e.key === 'Enter' && handleCreateCard()}
                 />
               </td>
               <td>
